chore(source/others/f32_eq_signaling.cpp) explain that `uint32_t` is not extended by `<<`

+++ source/others/f32_eq_signaling.cpp
@@ -47,6 +47,6 @@ f32_eq_signaling(float32_t a, float32_t b)
         softfloat_raiseFlags(softfloat_flag_invalid);
         return false;
     } else {
-        return uiA == uiB || 0 == ((uiA | uiB) << 1);
+        return uiA == uiB || 0 == static_cast<uint32_t>((uiA | uiB) << 1);
     }
 }
